+++ main-site/pages/[slug].celebPage.tsx
