+++ main-site/pages/[slug]/lp.tsx
