+++ main-site/pages/~kg/[kg].knowledgeGraphPage.tsx
